Fix clusterpools
